feat(shares): Phase 218 WebDAV 共享更新 API (PUT /api/v1/shares/webdav/{id})
- 使用 SqliteShareRepository 真实数据库实现
- JWT 认证，仅 admin 用户可访问
- 支持部分更新：name/path/description/public
- 验证共享 ID 存在性（404 Not Found）
- 验证协议为 WebDAV（非 WebDAV 返回 404）
- 验证名称/路径格式（400 Bad Request）
- 验证名称唯一性（409 Conflict，排除自身）
- 返回 200 OK + 更新后的共享详情
- 添加 docs/shares_webdav_update_api.md 文档

+++ src/handlers/shares_webdav_update.rs
@@ -120,7 +120,7 @@ pub async fn update_webdav_share(
     }
 
     // 6. 从数据库查询共享是否存在且为 WebDAV 协议
-    let existing_share = match repo.get_share_by_id(share_id as i64) {
+    let existing_share = match repo.get_share_by_id(share_id) {
         Ok(Some(s)) => {
             if s.protocol != "webdav" {
                 return Ok(HttpResponse::NotFound().json(ErrorResponse {
